Move AlertDialog creation to PrebookView
